+++ src/javascript/lib/core/patterns/defmatch.js
@@ -1,6 +1,7 @@
 /* @flow */
 
 import { buildMatch } from "./match";
+import * as Types from "./types";
 
 export class MatchError extends Error {
   constructor(arg: any) {
@@ -25,9 +26,13 @@ export class Case {
   pattern: Function;
   fn: Function;
   guard: Function;
+  arity: number;
+  optionals: Array<any>;
 
   constructor(pattern: Array<any>, fn: Function, guard: Function = () => true){
     this.pattern = buildMatch(pattern);
+    this.arity = pattern.length;
+    this.optionals = getOptionalValues(pattern);
     this.fn = fn;
     this.guard = guard;
   }
@@ -41,6 +46,8 @@ export function defmatch(...cases: Array<Case>): Function {
   return function(...args: Array<any>): any {
     for (let processedCase of cases) {
       let result = [];
+      args = fillInOptionalValues(args, processedCase.arity, processedCase.optionals);
+
       if (processedCase.pattern(args, result) && processedCase.guard.apply(this, result)) {
         return processedCase.fn.apply(this, result);
       }
@@ -48,6 +55,42 @@ export function defmatch(...cases: Array<Case>): Function {
 
     throw new MatchError(args);
   };
+}
+
+function getOptionalValues(pattern: Array<any>){
+  let optionals = [];
+
+  for(let i = 0; i < pattern.length; i++){
+    if(pattern[i] instanceof Types.Variable && pattern[i].default_value != Symbol.for("elixirscript.no_value")){
+      optionals.push([i, pattern[i].default_value]);
+    }
+  }
+
+  return optionals;
+}
+
+function fillInOptionalValues(args, arity, optionals){
+  if(args.length === arity || optionals.length === 0){
+    return args;
+  }
+
+  if(args.length + optionals.length < arity){
+    return args;
+  }
+
+  let numberOfOptionalsToFill = arity - args.length;
+  let optionalsToRemove = optionals.length - numberOfOptionalsToFill;
+
+  let optionalsToUse = optionals.slice(optionalsToRemove);
+
+  for(let [index, value] of optionalsToUse){
+    args.splice(index, 0, value);
+    if(args.length === arity){
+      break;
+    }
+  }
+
+  return args;
 }
 
 export function match(pattern: any, expr: any, guard: Function = () => true): Array<any> {